Move searchbar logic to searchbar.mjs
