+++ public/js/searchbar.mjs
@@ -1,0 +1,34 @@
+// Listener for searchbar
+
+// These are in pixel values for now
+// TODO Replace with actual data mapped through a SphereMercator transformation
+let tempFeatures = {
+	"pointa": { x: 3635, y: 919 },
+	"pointb": { x: 3365, y: 1288 }
+}
+
+function focusOn(camera, feature) {
+	if (tempFeatures[feature] == undefined) {
+		alert(`The feature ${feature} doesn't exist`);
+		return;
+	}
+	if (isNaN(tempFeatures[feature].x + tempFeatures[feature].y)) {
+		console.log(`Got NaN when looking for tempFeatures[${feature}] ${tempFeatures[feature]}`);
+		return;
+	}
+	camera.setCenterOn(tempFeatures[feature].x, tempFeatures[feature].y);
+	draw();
+}
+
+function addSearchbarListeners(camera) {
+	$("#searchbar-data").on("keyup", function (event) {
+		if (event.key == "Enter" || event.keyCode == 13) {
+			focusOn(camera, this.value);
+		}
+	});
+	$("#searchbar-submit").on("click", function (event) {
+		focusOn(camera, $("#searchbar-data")[0].value);
+	});
+}
+
+export { addSearchbarListeners };
